# apps/catalog-service/cmd/rest_api/docs/swagger.yaml
basePath: /catalog-service/v1
definitions:
  internal_infra_rest.categoryDTO:
    properties:
      category_id:
        description: O ID único da categoria.
        example: cat_2a1b3c4d5e
        type: string
      created_at:
        type: string
      description:
        description: A descrição da categoria.
        example: Eletrônicos
        type: string
      modified_at:
        type: string
    type: object
  internal_infra_rest.categoryInProductDTO:
    properties:
      category_id:
        description: ID público da categoria.
        example: cat_2a1b3c4d5e
        type: string
      description:
        description: Descrição da categoria.
        example: Eletrônicos
        type: string
    type: object
  internal_infra_rest.createCategoryDTO:
    properties:
      description:
        description: A descrição da categoria.
        example: Eletrônicos
        type: string
    required:
    - description
    type: object
  internal_infra_rest.createProductDTO:
    properties:
      category_id:
        description: ID público da Categoria à qual o produto pertence.
        example: cat_abc123def456
        type: string
      description:
        description: Descrição completa do produto.
        example: Smartphone X Pro 256GB
        type: string
      short_description:
        description: Descrição curta do produto.
        example: Smartphone X Pro
        type: string
      unit_of_measure:
        description: 'Unidade de medida (ex: UN, KG, MT).'
        example: UN
        type: string
    type: object
  internal_infra_rest.productDTO:
    properties:
      category:
        allOf:
        - $ref: '#/definitions/internal_infra_rest.categoryInProductDTO'
        description: Categoria associada ao produto.
      created_at:
        type: string
      description:
        description: Descrição completa do produto.
        example: Smartphone X Pro 256GB
        type: string
      modified_at:
        type: string
      product_id:
        description: ID público do produto.
        example: prod_xyz789uvw012
        type: string
      short_description:
        description: Descrição curta do produto.
        example: Smartphone X Pro
        type: string
      status:
        description: Status do produto (ACTIVE, INACTIVE).
        example: INACTIVE
        type: string
      unit_of_measure:
        description: 'Unidade de medida (ex: UN, KG, MT).'
        example: UN
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: contato@patrick.dev.br
    name: Patrick Ribeiro
    url: https://patrick.dev.br
  description: |-
    Serviço responsável por gerenciar o catálogo de produtos dentro do ERP.
    Faz parte da arquitetura de microserviços utilizada no portfólio de Patrick Ribeiro.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://github.com/openlab-software/erp
  title: Catalog Service API
  version: "1.0"
paths:
  /categories:
    get:
      description: Retorna uma lista de categorias. Permite filtrar por uma string
        de busca 'q'.
      parameters:
      - description: Termo de busca para filtrar categorias pela descrição
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de categorias
          schema:
            items:
              $ref: '#/definitions/internal_infra_rest.categoryDTO'
            type: array
      summary: List categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Cria uma nova categoria com base na descrição fornecida.
      parameters:
      - description: Payload para criação da categoria
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_infra_rest.createCategoryDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Categoria criada com sucesso
          schema:
            $ref: '#/definitions/internal_infra_rest.categoryDTO'
        "400":
          description: Requisição inválida (JSON mal formatado)
          schema:
            type: string
        "422":
          description: 'Erro de validação (ex: descrição em falta)'
          schema:
            type: string
      summary: Cria uma nova categoria
      tags:
      - Categories
  /categories/{id}:
    delete:
      description: Remove uma categoria específica do sistema pelo ID.
      parameters:
      - description: ID da Categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Categoria deletada com sucesso (sem conteúdo)
        "400":
          description: ID da categoria inválido
          schema:
            type: string
        "404":
          description: Categoria não encontrada
          schema:
            type: string
      summary: Deleta uma categoria
      tags:
      - Categories
    get:
      description: Retorna os detalhes de uma categoria específica.
      parameters:
      - description: ID da Categoria
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Detalhes da categoria
          schema:
            $ref: '#/definitions/internal_infra_rest.categoryDTO'
        "400":
          description: ID da categoria inválido
          schema:
            type: string
        "404":
          description: Categoria não encontrada
          schema:
            type: string
      summary: Busca uma categoria
      tags:
      - Categories
  /products:
    get:
      consumes:
      - application/json
      description: Retorna uma lista de produtos, opcionalmente filtrada por uma query
        de pesquisa 'q'.
      parameters:
      - description: 'Query de pesquisa (ex: descrição, id curto)'
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de produtos
          schema:
            items:
              $ref: '#/definitions/internal_infra_rest.productDTO'
            type: array
      summary: Lista todos os produtos
      tags:
      - products
    post:
      consumes:
      - application/json
      description: Cria um novo produto no catálogo com base no payload fornecido.
      parameters:
      - description: Payload para criação do produto
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/internal_infra_rest.createProductDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Produto criado com sucesso
          schema:
            $ref: '#/definitions/internal_infra_rest.productDTO'
      summary: Cria um novo produto
      tags:
      - products
  /products/{id}:
    delete:
      consumes:
      - application/json
      description: Deleta um produto permanentemente usando seu ID público.
      parameters:
      - description: ID público do produto (formato prod_...)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Deleta um produto
      tags:
      - products
    get:
      consumes:
      - application/json
      description: Retorna um único produto dado seu ID público.
      parameters:
      - description: ID público do produto (formato prod_...)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Produto encontrado
          schema:
            $ref: '#/definitions/internal_infra_rest.productDTO'
      summary: Obtém um produto pelo ID
      tags:
      - products
schemes:
- http
- https
swagger: "2.0"